+++ src/textual/__main__.py
@@ -11,7 +11,8 @@ if __name__ == "__main__":
             "[b magenta]Hope you liked the demo![/]\n\n"
             "Please consider sponsoring me if you get value from my work.\n\n"
             "Even the price of a ☕ can brighten my day!\n\n"
-            "https://github.com/sponsors/willmcgugan",
+            "https://github.com/sponsors/willmcgugan\n\n"
+            "- Will McGugan",
             border_style="red",
             title="Consider sponsoring",
         )