+++ src/textual/app.py
@@ -3985,7 +3985,6 @@ class App(Generic[ReturnType], DOMNode):
         action_target, action_name, params = self._parse_action(
             action, self if default_namespace is None else default_namespace
         )
-
         if action_target.check_action(action_name, params):
             return await self._dispatch_action(action_target, action_name, params)
         else: